refactor: remove models from utils-common
#190
