+++ libs/prices/auth/feature/src/lib/api/auth-credentials.model.ts
