+++ libs/prices/firestore/feature/src/lib/api/most-searched-item.model.ts
