+++ libs/prices/firestore/feature/src/lib/api/price.model.ts
@@ -1,0 +1,6 @@
+export interface Price {
+  productName: string;
+  price: number;
+  src: string;
+  location: string;
+}
