+++ libs/prices/home/feature/src/lib/api/most-searched-item.model.ts
@@ -1,0 +1,7 @@
+export interface MostSearchedItem {
+  id: string;
+  src: string;
+  name: string;
+  price: number;
+  location: string;
+}
