+++ libs/prices/store/feature/src/lib/api/auth-credentials.model.ts
@@ -1,0 +1,4 @@
+export interface AuthCredentials {
+  email: string;
+  password: string;
+}
