+++ libs/prices/store/feature/src/lib/mostSearched/adapter.ts
@@ -1,5 +1,5 @@
 import { createEntityAdapter } from '@ngrx/entity';
-import { MostSearchedItem } from '@travellers-apps/utils-common';
+import { MostSearchedItem } from '../api/most-searched-item.model';
 
 export const adapter = createEntityAdapter<MostSearchedItem>();
