+++ libs/utils-common/src/index.ts
@@ -1,6 +1,5 @@
 export { addNecessaryIcons } from './lib/add-necessary-icons';
 export * from './lib/get-ionic-config';
 export { Environment } from './lib/environment';
-export * from './lib/model';
 export * from './lib/consts';
 export { getMetaReducers } from './lib/get-meta-reducers';
