--- libs/utils-common/src/lib/model/index.ts
@@ -1,3 +1,0 @@
-export { MostSearchedItem } from './most-searched-item.model';
-export { AuthCredentials } from './auth-credentials.model';
-export { Price } from './price.model';
